Prepare upgrade to recent Quarkus version as part of reproducible builds
In Quarkus, I modified the JSON mapper to be more reproducible, we had
issues with the order of properties not being stable in the generated
JSON file.

This will require an update of the Quarkus version used, which I think
shouldn't be an issue.

There's one issue though: a constant has been dropped but AFAICS, we
changed to a YAML file in 0.27.0 so a veeeerrrryyyy long time ago.

Note: this will be a bit disruptive as the generated Platform will
probably contain a lot of changes compared to what was generated. But it
should be a lot more stable after that.